カードのパーツをコンポーネント化
