+++ frontend/src/components/molecules/CommonCardImg.vue
@@ -1,0 +1,18 @@
+<template>
+  <v-img
+    :aspect-ratio="88 / 63"
+    :src="imageUrl"
+    :alt="`${name}のカード`"
+  ></v-img>
+</template>
+
+<script>
+export default {
+  props: ['url', 'name'],
+  computed: {
+    imageUrl: function() {
+      return this.url ? this.url : '/img/no_cards_posted.jpg';
+    },
+  },
+};
+</script>
